content-generation-system/requirements.txt
python-dotenv==1.0.0
crewai==0.19.0
langchain>=0.1.10,<0.2.0
langchain_openai>=0.0.5
langchain_anthropic>=0.1.1
requests==2.31.0 